_vmconnect 1.1.0: enable starting a new ssh-agent.
By default, behavior is compatible with previous release. Set
`WITH_SSH_AGENT` to any non-empty string to spawn an agent for the
session.  This is used for the "is ssh up?" checks, and the final
connection to the VM.  If agent forwarding to the VM is enabled in your
config, then the VM will also be able to use the key in this agent.

Enabling `WITH_SSH_AGENT` prevents the SSH key from being stored to any
inherited `ssh-agent` process, such as the default agent running on OS
X.  The agent is killed (thus clearing its keys) when the connection
ends.