Delete vmconnect due to major VBox problems
Current VBoxManage stays connected to the shell which starts it somehow,
despite the use of `nohup ... &`, and POSIX sh does not give me enough
tools to daemonize it fully.

Exiting the shell causes the VM state to switch to `saved`.  Running a
shutdown command in the VM causes a race condition, where VirtualBox may
try to both save and power off the VM simultaneously, chewing on the CPU
for a long time.  Sometimes, it resolves itself; sometimes, the host
must be rebooted, and the VM becomes `aborted`.

Because these can't be cleaned up in shell (to my knowledge), and I
don't want to depend on anything else, the whole thing is unusable now.