Add `VMCONNECT_HOLD_ON_FAILURE` option
Improve usage in shell profiles where `_vmconnect` is "the command" and
exiting will close the window.  In such case, if the VM fails to start,
the window will disappear before the user can read the error messages.

Therefore, when `$VMCONNECT_HOLD_ON_FAILURE` is set and not 0, and
VBoxManage fails to launch the VM, the script waits for Enter to be
pressed before continuing.  (Implemented via `read`.)

If the environment variable is unset or zero, previous behavior applies;
the script will immediately exit upon failing to run VBoxManage.